# specs/001-farm-ops-platform/contracts/api.yaml
openapi: 3.1.0
info:
  title: FERM+ API
  version: 0.1.0
  description: >
    Contract for the initial FERM+ web platform covering authentication,
    farm management, intelligent agenda, alerts, and core operational modules.
servers:
  - url: https://api.ferm-plus.local
...
paths:
...
  /farms/{farmId}/inventory/items:
    get:
      tags: [Inventory]
      summary: List inventory items
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Inventory list
    post:
      tags: [Inventory]
      summary: Create an inventory item
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '201':
          description: Inventory item created
  /farms/{farmId}/sanitary/events:
    get:
      tags: [Sanitary]
      summary: List sanitary events
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Sanitary event list
    post:
      tags: [Sanitary]
      summary: Record a sanitary event
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '201':
          description: Sanitary event created
  /farms/{farmId}/finance/transactions:
    get:
      tags: [Finance]
      summary: List financial transactions
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '200':
          description: Transaction list
    post:
      tags: [Finance]
      summary: Record a financial transaction
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '201':
          description: Transaction created
...
  /farms/{farmId}/reports:
    post:
      tags: [Reports]
      summary: Generate a report
      parameters:
        - $ref: '#/components/parameters/FarmId'
      responses:
        '202':
          description: Report generation queued or completed
components:
  parameters:
    FarmId:
      in: path
      name: farmId
      required: true
      schema:
        type: string
...